feat: add `editor fullscreen` plugin (#14)
This is a pretty basic plugin, and an update to something you saw
earlier. It's an editor plugin, which Candle just added support for in
Bipsi. I don't know if you want to separate editor plugins from game
plugins in the list. It's not necessarily a binary designation, though I
suspect most plugins will either be "editor" or "game".

In case you aren't familiar with the editor plugin system:
- Any code following `//! CODE_RUNTIME`, or any code that doesn't foll
under one of these special comments (all plugins up till now) will be
handled like normal plugin code: added to the game in the editor's
playtest mode AND added to the game when exported.
- Any code following `//! CODE_EDITOR` is run ONLY in the editor: (1)
when the plugin is first added (2) when the app is first loaded and (3)
when a game is imported.
- Any code following `//! CODE_RUNTIME_DEV` is added to the game, but
only in the editor's playtest mode: it's NOT added to the game when
exported.

FYI, I've got 3 other editor plugins that are waiting on code-polish
before submitting to you.

I hope you don't mind all these new submissions I'm sending your way.

---------

Co-authored-by: Sean S. LeBlanc <sean.s.leblanc@gmail.com>