feat: add `smooth move` and `move animations` hacks (#27)
Candle recently weighed in on the code-injection technique used in this
plugin. They had no qualms with it, seeing it as no different from
rewriting the entire modified function in the plugin.
Also, I've split the plugin in two parts: smooth movement, and movement
animations. I've included both in this PR since they are so tied
together.

---------

Co-authored-by: Sean S. LeBlanc <sean.s.leblanc@gmail.com>